requirements.txt
# Core
textual==0.47.0
aiosqlite>=0.19.0
PyYAML>=6.0
aiohttp>=3.9.0

# LLM providers
openai>=1.0.0
anthropic>=0.25.0
google-generativeai>=0.5.0
groq>=0.4.0

# Development
pytest>=7.0.0
pytest-asyncio>=0.21.0
